refactor: DailyTask 기능을 Task로 통합하여 날짜 단위 태스크 관리로 변경
